# api/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "TAMS API",
    "description": "Time-addressable Media Store API - A comprehensive FastAPI implementation of the BBC TAMS API specification with VAST database integration for high-performance time-series analytics and S3-compatible storage for media segments.",
    "version": "6.0",
    "contact": {
      "name": "TAMS API Support",
      "url": "https://github.com/jesseVast/vasttams"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.tams.example.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/service": {
      "head": {
        "summary": "Return service path headers",
        "operationId": "head_service",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "summary": "Get service information",
        "operationId": "get_service",
        "tags": [
          "Service"
        ],
        "responses": {
          "200": {
            "description": "Service information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update service information",
        "operationId": "update_service",
        "tags": [
          "Service"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Service"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service information updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Service information updated"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Service": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "TAMS API"
          },
          "description": {
            "type": "string",
            "example": "Time-addressable Media Store API"
          },
          "type": {
            "type": "string",
            "example": "urn:x-tams:service:api"
          },
          "api_version": {
            "type": "string",
            "example": "6.0"
          },
          "service_version": {
            "type": "string",
            "example": "1.0.0"
          },
          "media_store": {
            "$ref": "#/components/schemas/MediaStore"
          },
          "event_stream_mechanisms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventStreamMechanism"
            }
          }
        }
      },
      "MediaStore": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "http_object_store"
          }
        }
      },
      "EventStreamMechanism": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "webhooks"
          },
          "description": {
            "type": "string",
            "example": "HTTP webhooks for event notifications"
          }
        }
      },
...
    }
  },
...
}